Update to Examples
I've changed the examples to be easier to just inherit from and get a working system off the bat.

PathGridObject.cs is now an abstract generic class. The generic defines what type the GenericGrid<T> sourceGrid is.

You'll find a new file in examples, this is a GridCell, which simply inherits from PathGridObject<GridCell> without any extras.

I've also removed the check for start node traversable in the Pathfinder.cs as I found it got in the way when I wanted to stop two different AIs being on the same GridCell.